requirements.txt
click==8.1.7
requests==2.32.3
python-dotenv==1.0.1
pendulum==3.0.0
tabulate==0.9.0
tenacity==9.0.0
rich==13.9.4
schedule==1.2.0
flask==3.0.0
flask-cors==4.0.0

